move info service commands into orderbook service (#407)
* move info service commands into orderbook service

* update commands

* rename folder from info to market

* use market instead of info in cli, update spec

* fix doc
