+++ broker-cli/commands/market/index.js
@@ -2,12 +2,12 @@ const { validations } = require('../../utils')
 const { RPC_ADDRESS_HELP_STRING, MARKET_NAME_HELP_STRING } = require('../../utils/strings')
 
 /**
- * Info
- * @module broker-cli/info
+ * Market
+ * @module broker-cli/market
  */
 
 /**
- * Supported commands for `sparkswap info`
+ * Supported commands for `sparkswap market`
  *
  * @constant
  * @type {Object<key, String>}
@@ -25,7 +25,7 @@ const trades = require('./trades')
 
 module.exports = (program) => {
   program
-    .command('info', 'Commands to get market, trading, and fee information')
+    .command('market', 'Commands to get market, trading, and fee information')
     .help(`Available Commands: ${Object.values(SUPPORTED_COMMANDS).join(', ')}`)
     .argument('<command>', '', Object.values(SUPPORTED_COMMANDS), null, true)
     .argument('[sub-arguments...]')
@@ -50,11 +50,11 @@ module.exports = (program) => {
           return trades(args, opts, logger)
       }
     })
-    .command(`info ${SUPPORTED_COMMANDS.SUPPORTED_MARKETS}`, 'Get the markets currently supported')
+    .command(`market ${SUPPORTED_COMMANDS.SUPPORTED_MARKETS}`, 'Get the markets currently supported')
     .option('--rpc-address [rpc-address]', RPC_ADDRESS_HELP_STRING, validations.isHost)
-    .command(`info ${SUPPORTED_COMMANDS.MARKET_STATS}`, 'Get statistics (price ticker information) for a particular market for a period of the last 24 hours')
+    .command(`market ${SUPPORTED_COMMANDS.MARKET_STATS}`, 'Get statistics (price ticker information) for a particular market for a period of the last 24 hours')
     .option('--market [marketName]', MARKET_NAME_HELP_STRING, validations.isMarketName)
-    .command(`info ${SUPPORTED_COMMANDS.TRADES}`, 'Get detailed information about trades from a given time range')
+    .command(`market ${SUPPORTED_COMMANDS.TRADES}`, 'Get detailed information about trades from a given time range')
     .argument('<since>', 'Start datetime in ISO format e.g. 2018-04-23T10:26:00.996Z', validations.isDate)
     .argument('<limit>', 'Number of records to be retrieved', validations.isPositiveInteger)
     .option('--rpc-address [rpc-address]', RPC_ADDRESS_HELP_STRING, validations.isHost)
