+++ broker-daemon/broker-rpc/broker-rpc-server.spec.js
@@ -14,8 +14,6 @@ describe('BrokerRPCServer', () => {
   let orderBookService
   let WalletService
   let walletService
-  let InfoService
-  let infoService
   let pathResolve
   let protoPath
   let engines
@@ -52,13 +50,6 @@ describe('BrokerRPCServer', () => {
     }
     WalletService = sinon.stub().returns(walletService)
     BrokerRPCServer.__set__('WalletService', WalletService)
-
-    infoService = {
-      definition: 'mydef',
-      implementation: 'myimp'
-    }
-    InfoService = sinon.stub().returns(infoService)
-    BrokerRPCServer.__set__('InfoService', InfoService)
 
     addService = sinon.stub()
     rpcServer = sinon.stub().returns({
@@ -229,27 +220,6 @@ describe('BrokerRPCServer', () => {
       expect(WalletService).to.have.been.calledWithNew()
       expect(server).to.have.property('walletService')
       expect(server.walletService).to.be.equal(walletService)
-    })
-
-    it('adds the info service', () => {
-      const server = new BrokerRPCServer()
-
-      expect(server).to.have.property('server')
-      expect(server.server.addService).to.be.equal(addService)
-      expect(addService).to.have.been.calledWith(infoService.definition, infoService.implementation)
-    })
-
-    it('creates a info service', () => {
-      const logger = 'mylogger'
-      const engines = 'myengines'
-
-      const server = new BrokerRPCServer({ logger, engines })
-
-      expect(InfoService).to.have.been.calledOnce()
-      expect(InfoService).to.have.been.calledWith(protoPath, sinon.match({ logger, engines }))
-      expect(InfoService).to.have.been.calledWithNew()
-      expect(server).to.have.property('infoService')
-      expect(server.infoService).to.be.equal(infoService)
     })
 
     it('adds the orderBook service', () => {
