+++ broker-daemon/broker-rpc/orderbook-service/index.spec.js
@@ -6,6 +6,9 @@ const OrderBookService = rewire(path.resolve(__dirname))
 describe('OrderBookService', () => {
   let watchMarketStub
   let getOrderbookStub
+  let getSupportedMarketsStub
+  let getMarketStatsStub
+  let getTradesStub
 
   let GrpcMethod
   let register
@@ -30,7 +33,10 @@ describe('OrderBookService', () => {
             service: 'fakeService'
           },
           WatchMarketResponse: sinon.stub(),
-          GetOrderbookResponse: sinon.stub()
+          GetOrderbookResponse: sinon.stub(),
+          GetSupportedMarketsResponse: sinon.stub(),
+          GetMarketStatsResponse: sinon.stub(),
+          GetTradesResponse: sinon.stub()
         }
       }
     }
@@ -59,6 +65,15 @@ describe('OrderBookService', () => {
 
     getOrderbookStub = sinon.stub()
     OrderBookService.__set__('getOrderbook', getOrderbookStub)
+
+    getSupportedMarketsStub = sinon.stub()
+    OrderBookService.__set__('getSupportedMarkets', getSupportedMarketsStub)
+
+    getMarketStatsStub = sinon.stub()
+    OrderBookService.__set__('getMarketStats', getMarketStatsStub)
+
+    getTradesStub = sinon.stub()
+    OrderBookService.__set__('getTrades', getTradesStub)
   })
 
   beforeEach(() => {
@@ -194,6 +209,146 @@ describe('OrderBookService', () => {
 
     it('passes in the response', () => {
       expect(callArgs[3]).to.be.eql({ GetOrderbookResponse: proto.broker.rpc.GetOrderbookResponse })
+    })
+  })
+
+  describe('#getSupportedMarkets', () => {
+    let callOrder = 2
+    let callArgs
+
+    beforeEach(() => {
+      callArgs = GrpcMethod.args[callOrder]
+    })
+
+    it('exposes an implementation', () => {
+      expect(server.implementation).to.have.property('getSupportedMarkets')
+      expect(server.implementation.getSupportedMarkets).to.be.a('function')
+    })
+
+    it('creates a GrpcMethod', () => {
+      expect(GrpcMethod).to.have.been.called()
+      expect(GrpcMethod).to.have.been.calledWithNew()
+      expect(server.implementation.getSupportedMarkets).to.be.equal(fakeRegistered)
+    })
+
+    it('provides the method', () => {
+      expect(callArgs[0]).to.be.equal(getSupportedMarketsStub)
+    })
+
+    it('provides a message id', () => {
+      expect(callArgs[1]).to.be.equal('[OrderBookService:getSupportedMarkets]')
+    })
+
+    describe('request options', () => {
+      it('passes in the logger', () => {
+        expect(callArgs[2]).to.have.property('logger', logger)
+        expect(callArgs[2].logger).to.be.equal(logger)
+      })
+
+      it('passes in the relayer', () => {
+        expect(callArgs[2]).to.have.property('relayer')
+        expect(callArgs[2].relayer).to.be.equal(relayer)
+      })
+
+      it('passes in the orderbooks', () => {
+        expect(callArgs[2]).to.have.property('orderbooks')
+        expect(callArgs[2].orderbooks).to.be.equal(orderbooks)
+      })
+    })
+
+    it('passes in the response', () => {
+      expect(callArgs[3]).to.be.an('object')
+      expect(callArgs[3]).to.have.property('GetSupportedMarketsResponse', proto.broker.rpc.GetSupportedMarketsResponse)
+    })
+  })
+
+  describe('#getMarketStatsMarkets', () => {
+    let callOrder = 3
+    let callArgs
+
+    beforeEach(() => {
+      callArgs = GrpcMethod.args[callOrder]
+    })
+
+    it('exposes an implementation', () => {
+      expect(server.implementation).to.have.property('getMarketStats')
+      expect(server.implementation.getMarketStats).to.be.a('function')
+    })
+
+    it('creates a GrpcMethod', () => {
+      expect(GrpcMethod).to.have.been.called()
+      expect(GrpcMethod).to.have.been.calledWithNew()
+      expect(server.implementation.getMarketStats).to.be.equal(fakeRegistered)
+    })
+
+    it('provides the method', () => {
+      expect(callArgs[0]).to.be.equal(getMarketStatsStub)
+    })
+
+    it('provides a message id', () => {
+      expect(callArgs[1]).to.be.equal('[OrderBookService:getMarketStats]')
+    })
+
+    describe('request options', () => {
+      it('passes in the logger', () => {
+        expect(callArgs[2]).to.have.property('logger', logger)
+        expect(callArgs[2].logger).to.be.equal(logger)
+      })
+
+      it('passes in the orderbooks', () => {
+        expect(callArgs[2]).to.have.property('orderbooks')
+        expect(callArgs[2].orderbooks).to.be.equal(orderbooks)
+      })
+    })
+
+    it('passes in the response', () => {
+      expect(callArgs[3]).to.be.an('object')
+      expect(callArgs[3]).to.have.property('GetMarketStatsResponse', proto.broker.rpc.GetMarketStatsResponse)
+    })
+  })
+
+  describe('#getTrades', () => {
+    let callOrder = 4
+    let callArgs
+
+    beforeEach(() => {
+      callArgs = GrpcMethod.args[callOrder]
+    })
+
+    it('exposes an implementation', () => {
+      expect(server.implementation).to.have.property('getTrades')
+      expect(server.implementation.getTrades).to.be.a('function')
+    })
+
+    it('creates a GrpcMethod', () => {
+      expect(GrpcMethod).to.have.been.called()
+      expect(GrpcMethod).to.have.been.calledWithNew()
+      expect(server.implementation.getTrades).to.be.equal(fakeRegistered)
+    })
+
+    it('provides the method', () => {
+      expect(callArgs[0]).to.be.equal(getTradesStub)
+    })
+
+    it('provides a message id', () => {
+      expect(callArgs[1]).to.be.equal('[OrderBookService:getTrades]')
+    })
+
+    describe('request options', () => {
+      it('passes in the logger', () => {
+        expect(callArgs[2]).to.have.property('logger', logger)
+        expect(callArgs[2].logger).to.be.equal(logger)
+      })
+
+      it('passes in the orderbooks', () => {
+        expect(callArgs[2]).to.have.property('orderbooks')
+        expect(callArgs[2].orderbooks).to.be.equal(orderbooks)
+      })
+    })
+
+    it('passes in the response', () => {
+      expect(callArgs[3]).to.be.an('object')
+      expect(callArgs[3]).to.have.property('GetTradesResponse', proto.broker.rpc.GetTradesResponse)
     })
   })
 })
